[refactor] Move complex hooks and logics into new files
